Initial version.
This is a UDI Polyglot node server that listens for data published
by the Weather Display software.

It is based on the WeatherFlow node server and attempts to map
Weather Display data to WeatherFlow data, this means there are missing
data fields and the units have not been verified.

This is experimental.