Parsing custom parameters for data value to WD field mapping.
We want to allow the user to configure what data fields are actually
used from the WD data. Sending all 57+ fields to the ISY is not ideal.

Parse custom parameters that are formatted as <node>-<driver> with the
WD field number as the value. This will generate lists of drivers for
each node type.  Using that we can create custom node definitions that
contain only the values the user is interested in.

The mapping will also need to take into account the units so that the
correct editor can be configured for each driver.